[Modify]: 파일&경로 convention  통합 및 ERD 추가 수정
