+++ src/controllers/index.js
@@ -1,5 +1,7 @@
 const productController = require('./productController');
+const kakaoController = require('./kakaoControllers')
 
 module.exports = {
   productController,
+  kakaoController
 };
