+++ src/controllers/productController.js
@@ -12,8 +12,41 @@ const getListByType = catchAsync(async (req, res, next) => {
   return res.status(200).json(list);
 }
 )
+const getProductByName = async (req, res) => {
+  try {
+    const { productName } = req.query;
+    console.log(req.query)
+    if (!productName) {
+      throw new Error("Key Error");
+    }
+
+    const result = await productService.getProductByName(productName);
+
+    return res.status(200).json(result);
+  } catch (err) {
+    res.status(err.statusCode || 400).json({ message: err.message });
+  }
+};
+
+
+const getDetailByProductId = async (req, res) => {
+  try {
+    const productId = req.params.productId;
+
+    if (!productId) {
+      throw new Error("Key Error");
+    }
+    const result = await productService.getDetailByProductId(productId);
+
+    return res.status(200).json(result);
+  } catch (err) {
+    res.status(err.statusCode || 400).json({ message: err.message });
+  }
+};
 
 module.exports = {
   getListByType,
-  getAllProducts
+  getAllProducts,
+  getProductByName,
+  getDetailByProductId
 };
