--- src/routes/product.router.js
@@ -1,10 +1,0 @@
-const express = require("express");
-
-const productController = require("../controllers/product.controller");
-
-const productRouter = express.Router();
-
-productRouter.get("/detail/:productId", productController.getDetailByProductId);
-productRouter.get("/name", productController.getProductByName);
-
-module.exports = { productRouter };
