+++ src/routes/productRouter.js
@@ -5,5 +5,7 @@ const router = express.Router();
 
 router.get('/', productController.getAllProducts);
 router.get('/listing', productController.getListByType);
+router.get("/detail/:productId", productController.getDetailByProductId);
+router.get("/name", productController.getProductByName);
 
 module.exports = router;
