+++ src/service/index.js
@@ -1,5 +1,7 @@
 const productService = require('./productService');
+const kakaoService = require('./kakaoService')
 
 module.exports = {
   productService,
+  kakaoService
 };
