--- src/service/product.service.js
@@ -1,7 +1,0 @@
-const productDao = require("../models/product.dao");
-
-const getDetailByProductId = async (productId) => {
-  return productDao.getDetailByProductId(productId);
-};
-
-module.exports = { getDetailByProductId };
