--- src/services/product.service.js
@@ -1,7 +1,0 @@
-const productDao = require("../models/product.dao");
-
-const getProductByName = async (productName) => {
-  return productDao.getProductByName(productName);
-};
-
-module.exports = { getProductByName };
